+++ src/OpenCensus.Exporter.Zipkin/ZipkinTraceExporterOptions.cs
@@ -27,7 +27,7 @@ namespace OpenCensus.Exporter.Zipkin
         /// Gets or sets Zipkin endpoint address. See https://zipkin.io/zipkin-api/#/default/post_spans. 
         /// Typically https://zipkin-server-name:9411/api/v2/spans.
         /// </summary>
-        public string Endpoint { get; set; } = "http://localhost:9411/api/v2/spans";
+        public Uri Endpoint { get; set; } = new Uri("http://localhost:9411/api/v2/spans");
 
         /// <summary>
         /// Gets or sets timeout in seconds.