+++ src/OpenCensus/Impl/Trace/Export/LatencyBucketBoundaries.cs
@@ -32,7 +32,7 @@ namespace OpenCensus.Trace.Export
 
         public static IList<ISampledLatencyBucketBoundaries> Values = new List<ISampledLatencyBucketBoundaries>()
         {
-            ZERO_MICROSx10, MICROSx10_MICROSx100, MICROSx100_MILLIx1, MILLIx1_MILLIx10, MILLIx10_MILLIx100, MILLIx100_SECONDx1, SECONDx1_SECONDx10, SECONDx10_SECONDx100, SECONDx100_MAX
+            ZERO_MICROSx10, MICROSx10_MICROSx100, MICROSx100_MILLIx1, MILLIx1_MILLIx10, MILLIx10_MILLIx100, MILLIx100_SECONDx1, SECONDx1_SECONDx10, SECONDx10_SECONDx100, SECONDx100_MAX,
         };
 
         internal LatencyBucketBoundaries(long latencyLowerNs, long latencyUpperNs)