+++ src/OpenCensus/Impl/Trace/LinkType.cs
@@ -19,6 +19,6 @@ namespace OpenCensus.Trace
     public enum LinkType
     {
         CHILD_LINKED_SPAN,
-        PARENT_LINKED_SPAN
+        PARENT_LINKED_SPAN,
     }
 }