+++ src/OpenCensus/Impl/Trace/MessageEventType.cs
@@ -19,6 +19,6 @@ namespace OpenCensus.Trace
     public enum MessageEventType
     {
         SENT,
-        RECEIVED
+        RECEIVED,
     }
 }