+++ test/OpenCensus.Tests/Impl/Trace/Export/InProcessSampledSpanStoreTest.cs
@@ -48,7 +48,7 @@ namespace OpenCensus.Trace.Export.Test
         public InProcessSampledSpanStoreTest()
         {
             sampledSpanContext = SpanContext.Create(TraceId.GenerateRandomId(random), SpanId.GenerateRandomId(random), TraceOptions.Builder().SetIsSampled(true).Build());
-            notSampledSpanContext = SpanContext.Create(TraceId.GenerateRandomId(random), SpanId.GenerateRandomId(random), TraceOptions.DEFAULT);
+            notSampledSpanContext = SpanContext.Create(TraceId.GenerateRandomId(random), SpanId.GenerateRandomId(random), TraceOptions.Default);
             parentSpanId = SpanId.GenerateRandomId(random);
             startEndHandler = new TestStartEndHandler(sampleStore);
             sampleStore.RegisterSpanNamesForCollection(new List<string>() { REGISTERED_SPAN_NAME });