+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Initial version of Application Insights exporter implemented.
 - Zipkin exporter implemented.
 - Initial version of SDK published. It is based on contribution from Pivotal
   [from](https://github.com/SteeltoeOSS/Management/tree/dev/src/Steeltoe.Management.OpenCensus).